DTSCCI=958 Refactor MediationJsonService
