+++ src/main/java/uk/gov/hmcts/reform/civil/service/mediation/helpers/PartyDetailPopulator.java
@@ -1,0 +1,23 @@
+package uk.gov.hmcts.reform.civil.service.mediation.helpers;
+
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.model.Party;
+import uk.gov.hmcts.reform.civil.service.mediation.MediationLitigant;
+
+@Component
+public class PartyDetailPopulator {
+
+    private static final String PAPER_RESPONSE = "N";
+
+    public MediationLitigant.MediationLitigantBuilder populator(MediationLitigant.MediationLitigantBuilder builder,
+                                                                Party party) {
+        String partyRole = party.getFlags() != null ? party.getFlags().getRoleOnCase() : null;
+        return builder.partyID(party.getPartyID())
+            .partyRole(partyRole)
+            .partyType(party.getType())
+            .partyName(party.getPartyName())
+            .paperResponse(PAPER_RESPONSE);
+    }
+
+}
